pyproject.toml
[project]
name = "pptx-mcp-server"
version = "1.0.0"
description = "MCP Server for PowerPoint presentation creation, editing, and analysis"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "python-pptx>=0.6.21",
    "Pillow>=10.0.0",
    "lxml>=4.9.0",
    "defusedxml>=0.7.1",
    "six>=1.16.0",
]

[project.scripts]
pptx-mcp-server = "pptx_mcp_server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["pptx_mcp_server"]
